Extracts vehicle concept into a separate class
